# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DS ESTIMATION BACKEND DOCUMENTATION",
    "description": "Our story started in 2015 when there was a need of Construction database for helping bidders to bid accurately with reliable information's and estimates in the Rwanda construction industry. Continuous determination, hard work and passion have been constant themes throughout our history and present. As Rwanda Construction industry becomes increasingly important on the global stage, it's our believe that, with our own innovative mindset, we can make the construction industry a better one for national development. So we challenge ourselves to evolve, innovate and experiment to make work easier and faster for everyone as the leading local software brand in the Rwanda construction industry. We are currently the quantities and estimations software dealers which are made in Africa its self initially from Ghana and same technology has been customized to fit Rwanda construction industry and we are super proud to start the race with national development in minds. Today we are privileged to connect with thousands of Engineers and other construction related professionals who do daily estimates or quantities in the construction industry.",
    "version": "1.0.0",
    "contact": {
      "name": "DS ESTIMATION",
      "url": "http://dsestimation.com",
      "email": "dsestimation@gmail.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "AuthToken": {
      "type": "apiKey",
      "name": "auth-token",
      "in": "header",
      "description": "The token for authentication"
    }
  },
  "security": [
    {
      "AuthToken": []
    }
  ],
  "paths": {
...
    "/owner/subcontractor/create": {
      "post": {
        "tags": ["Default Subcontractor Routes"],
        "summary": "Create a subcontractor",
        "description": "Create a subcontractor",
        "operationId": "createSubcontractor",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Subcontractor object that needs to be created subcontractor",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Subcontractor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcontractor created successfully"
          },
          "400": {
            "description": "Some subcontractor field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/subcontractor/{id}/update": {
      "patch": {
        "tags": ["Default Subcontractor Routes"],
        "summary": "Update a subcontractor",
        "description": "Update a subcontractor",
        "operationId": "updateSubcontractor",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the subcontractor to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Subcontractor object that needs to be updated default subcontractor, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Subcontractor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcontractor updated successfully"
          },
          "400": {
            "description": "Some subcontractor field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Subcontractor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/subcontractor/{id}/delete": {
      "delete": {
        "tags": ["Default Subcontractor Routes"],
        "summary": "Delete a subcontractor",
        "description": "Delete a subcontractor",
        "operationId": "deleteSubcontractor",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the subcontractor to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Subcontractor deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Subcontractor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/supplier/all": {
      "get": {
        "tags": ["Client Supplier Routes"],
        "summary": "Get all suppliers",
        "description": "Get all suppliers",
        "operationId": "getAllSuppliers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Suppliers Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/supplier/create": {
      "post": {
        "tags": ["Client Supplier Routes"],
        "summary": "Create a supplier",
        "description": "Create a supplier",
        "operationId": "createSupplier",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Supplier object that needs to be created supplier",
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier created successfully"
          },
          "400": {
            "description": "Some supplier field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/supplier/{id}/update": {
      "patch": {
        "tags": ["Client Supplier Routes"],
        "summary": "Update a supplier",
        "description": "Update a supplier",
        "operationId": "updateSupplier",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the supplier to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Supplier object that needs to be updated supplier",
            "schema": {
              "$ref": "#/definitions/Supplier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier updated successfully"
          },
          "400": {
            "description": "Some supplier field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Supplier not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/supplier/{id}/delete": {
      "delete": {
        "tags": ["Client Supplier Routes"],
        "summary": "Delete a supplier",
        "description": "Delete a supplier",
        "operationId": "deleteSupplier",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the supplier to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Supplier deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Supplier not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/subcontractor/all": {
      "get": {
        "tags": ["Client Subcontractor Routes"],
        "summary": "Get all subcontractors",
        "description": "Get all subcontractors",
        "operationId": "getAllSubcontractors",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Subcontractors Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/subcontractor/create": {
      "post": {
        "tags": ["Client Subcontractor Routes"],
        "summary": "Create a subcontractor",
        "description": "Create a subcontractor",
        "operationId": "createSubcontractor",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Subcontractor object that needs to be created subcontractor",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Subcontractor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcontractor created successfully"
          },
          "400": {
            "description": "Some subcontractor field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/subcontractor/{id}/update": {
      "patch": {
        "tags": ["Client Subcontractor Routes"],
        "summary": "Update a subcontractor",
        "description": "Update a subcontractor",
        "operationId": "updateSubcontractor",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the subcontractor to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Subcontractor object that needs to be updated Client subcontractor, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Subcontractor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcontractor updated successfully"
          },
          "400": {
            "description": "Some subcontractor field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Subcontractor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/subcontractor/{id}/delete": {
      "delete": {
        "tags": ["Client Subcontractor Routes"],
        "summary": "Delete a subcontractor",
        "description": "Delete a subcontractor",
        "operationId": "deleteSubcontractor",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the subcontractor to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Subcontractor deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Subcontractor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/global/subcontractor/all": {
      "get": {
        "tags": ["Global Routes"],
        "summary": "Get all subcontractors from our database",
        "description": "Get all subcontractors from our database",
        "operationId": "getAllSubcontractors",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Subcontractors Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/global/supplier/all": {
      "get": {
        "tags": ["Global Routes"],
        "summary": "Get all suppliers from our database",
        "description": "Get all suppliers from our database",
        "operationId": "getAllSuppliers",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Suppliers Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Supplier": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "projectId": {
          "type": "number"
        }
      }
    },
...
  }
}